// src/features/auth/schemas.ts
import {z} from 'zod';

export const loginSchema = z.object({
  email: z.string().email().max(255),
  password: z.string().min(1).max(100),
});

export const registerSchema = z
  .object({
    name: z.string().min(2).max(80),
    email: z.string().email().max(255),
    password: z.string().min(8).max(100),
    confirmPassword: z.string(),
  })
  .refine((data) => data.password === data.confirmPassword, {
    path: ['confirmPassword'],
  });